test: split the public window spec by concern
