+++ e2e/helpers/public-window-reading-fixtures.ts
@@ -1,0 +1,5 @@
+export const NOTE_EXCERPT = 'The public note begins here'
+
+export const THING_EXCERPT = 'A lantern with an abbreviated inscription'
+
+export const NOTE_FULL = `${NOTE_EXCERPT}, then continues beyond the snapshot excerpt.`
